Always Make Sure the NGINX Service Is Started
There are some situations where the NGINX service is not started (the
start service is skipped because nginx_config.stat.exists is
true--because the main NGINX config file already exists) but the restart
handler is fired. In such cases the handler fails because it expects
NGINX to have already been running.

Make the task checkig if NGINX is running non-conditional to avoid these
situation.

Signed-off-by: Jason Rogena <jasonrogena@gmail.com>