Fix NGINX (Re)start Issue
Fixes the issue causing Ansible to fail with a variable undefined error
for the "nginx_first_start" variable.

Simplifies the process by allowing both the start and restart handlers
to run in the same build. The start handler will, however, always run
before the restart handler (based on the placement in the handlers/main.yml
file).

Signed-off-by: Jason Rogena <jasonrogena@gmail.com>