Switch to Molecule For Tests
Switch to molecule (https://molecule.readthedocs.io) for testing the
role. Molecule is a testing framework by Ansible. It supports Vagrant,
so we should be able to port the tests we were previously running directly
in Vagrant to it.

Molecule supports linting (Ansible and YAML) and infrastructure testing
(using a tool like testinfra) out of the box.

Signed-off-by: Jason Rogena <jasonrogena@gmail.com>