Move the access_log_format Variable
Since the access_log_format variable is only used in sites that have SSL
turned on, moved it to inside the ssl map of sites.

This will prevent it from being directly placed in a site file and
causing the NGINX service to error with "an unknown directive
access_log_format".

Signed-off-by: Jason Rogena <jasonrogena@gmail.com>