Reload NGINX after (re)start fires
In cases where both the reload and either the start or restart handlers
are scheduled to fire, perform the reload task after the start and
restart handlers.

This prevents issues where the reload handler fails with an error
suggesting the PID file doesn't exist because the NGINX service isn't
running yet.

Signed-off-by: Jason Rogena <jason@rogena.me>